simplify plugin
